[Refactor] Further clean up
